backend/requirements.txt
pydub
supabase
uvicorn
fastapi[all]
uvicorn[standard]
python-dotenv
requests
openai
opencv-python
pdf2image
gtts
tesseract
openai
python-multipart
moviepy==1.0.3
slowapi
ratelimit
anyio